insert error code checking

+++ shell/insert.sh
@@ -24,6 +24,11 @@ mkdir -p ../tmp
 
 cat ../sql/templates/insert_user | sed -e "s/<username>/$username/g" -e "s/<password>/$password/g" > ../tmp/insert_user.sql
 
-psql -U postgres -f ../tmp/insert_user.sql -d postgres
+psql -U postgres -f ../tmp/insert_user.sql -d postgres -v "ON_ERROR_STOP=1"
+
+if [ $? -ne 0 ];
+then
+    exit $?
+fi
 
 rm ../tmp/insert_user.sql
